requirements.txt
python-dotenv
openai
fastapi
uvicorn
termcolor
aiofiles
python-multipart
fpdf 